prevent poll closing on poll pane hide
